added change pin menu & eeprom
testing add user

+++ inc/eeprom.h
@@ -19,6 +19,6 @@ EEPROM_STATUS getBootStamp(RTC_TIME_T* boottime);
 EEPROM_STATUS getUserData(uint8_t userid, struct users_S *userdata);
 EEPROM_STATUS addUser(struct users_S *userdata);
 EEPROM_STATUS saveNewUser(uint8_t userid, struct users_S *newuser);
-EEPROM_STATUS changePIN(uint8_t userid, uint8_t *newpin);
+EEPROM_STATUS changePIN(uint8_t userid, uint32_t *newpin);
 
 #endif /* EEPROM_H_ */
